able to add cuisine after adding name from the frontend to the backend
