+++ src/listContainer.jsx
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from "react";
 import { hot } from 'react-hot-loader/root';
 import FoodOption from "./foodOption";
 
-function ListContainer() {
+function ListContainer(props) {
+    const { user_id } = props
     const tally = {};
     const options = ['Japanese', 'Mexican', 'Thai', 'Korean'];
 
@@ -31,26 +32,11 @@ function ListContainer() {
 
 
     const upvoteFunc = async (selection) => {
-        // try {
-        //     await fetch('http://localhost:3000/api', {
-        //         mode: 'no-cors',
-        //         method: 'POST',
-        //         headers: {
-        //             'Content-Type':'application/json'
-        //         },
-        //         body: JSON.stringify({name:'KingDANG'})
-        //     });
-        //     if (response.ok) {
-        //         console.log('added to db')
-        //     }
-        // }catch (error){
-        //     console.log('Error in upvoteFunc ', error)
-        // };
         setVote(selection);
         if(selection !== 'nobody yet'){
             tally[selection] = tally[selection] + 1;
             setTallyData({...tally});
-            fetchData(); 
+            // fetchData(); 
             //UNCOMMENT AFTER ACTUALLY BEING ABLE TO FETCH DATA
         };
     };
@@ -58,7 +44,7 @@ function ListContainer() {
     const list = [];
     options.forEach(el => {
         list.push(
-            <FoodOption key={el} value={el} currVote={vote} upvoteFunc={upvoteFunc}/>
+            <FoodOption key={el} value={el} currVote={vote} upvoteFunc={upvoteFunc} user_id={user_id}/>
             );
         }
     );
